Move commune demo to commune folder
